test(chat): split chat command tests
