+++ src/chat-app.test.ts
@@ -1,5 +1,4 @@
 import { describe, expect, test } from "bun:test";
-import { formatSessionList, resolveResumeSession } from "./chat-commands";
 import {
   applyAtSuggestion,
   extractAtReferencePaths,
@@ -9,49 +8,9 @@ import {
 } from "./chat-file-ref";
 import { appendPromotedItems, applyPromotion } from "./chat-promotion";
 import { toRows } from "./chat-session";
-import { createSession, createStore } from "./test-utils";
-
-function createUiStore() {
-  return createStore({
-    activeSessionId: "sess_aaaa1111",
-    sessions: [
-      createSession({ id: "sess_aaaa1111", title: "First" }),
-      createSession({ id: "sess_bbbb2222", title: "Second" }),
-    ],
-  });
-}
+import { createSession } from "./test-utils";
 
 describe("chat-ui helpers", () => {
-  test("resolveResumeSession reports usage when no prefix is provided", () => {
-    const resolved = resolveResumeSession(createUiStore(), "/resume");
-    expect(resolved.kind).toBe("usage");
-  });
-
-  test("resolveResumeSession reports not_found for unknown prefix", () => {
-    const resolved = resolveResumeSession(createUiStore(), "/resume sess_missing");
-    expect(resolved.kind).toBe("not_found");
-    if (resolved.kind === "not_found") expect(resolved.prefix).toBe("sess_missing");
-  });
-
-  test("resolveResumeSession reports ambiguous for multi-match prefix", () => {
-    const store = createUiStore();
-    const resolved = resolveResumeSession(store, "/resume sess_");
-    expect(resolved.kind).toBe("ambiguous");
-    if (resolved.kind === "ambiguous") expect(resolved.matches).toHaveLength(2);
-  });
-
-  test("resolveResumeSession returns target session for exact-ish prefix", () => {
-    const resolved = resolveResumeSession(createUiStore(), "/resume sess_bbbb");
-    expect(resolved.kind).toBe("ok");
-    if (resolved.kind === "ok") expect(resolved.session.id).toBe("sess_bbbb2222");
-  });
-
-  test("formatSessionList marks active session", () => {
-    const lines = formatSessionList(createUiStore());
-    expect(lines[0]?.startsWith("● ")).toBe(true);
-    expect(lines[1]?.startsWith("  ")).toBe(true);
-  });
-
   test("extractAtReferenceQuery parses @prefix", () => {
     expect(extractAtReferenceQuery("@cli")).toBe("cli");
     expect(extractAtReferenceQuery(" @cli")).toBe("cli");
